# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: ClawChain API
  version: 0.1.0
  description: REST API for the ClawChain AI-Native Blockchain
  contact:
    name: ClawChain Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:1317
    description: Local development node
  - url: https://api-testnet.clawchain.io
    description: ClawChain Testnet
...
paths:
...
  /faucet:
    post:
      tags: [Faucet]
      summary: Request testnet tokens
      description: Sends testnet CLAW tokens to the provided address. Rate-limited per address.
      operationId: requestFaucet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - address
              properties:
                address:
                  type: string
                  description: Recipient bech32 address
                  example: "claw1abc123..."
      responses:
        "200":
          description: Tokens sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  txHash:
                    type: string
              example:
                message: "Sent 10 CLAW to claw1abc123..."
                txHash: "A1B2C3..."
        "400":
          description: Invalid address or rate-limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: "Rate limit exceeded. Try again in 60 seconds."
        "500":
          $ref: "#/components/responses/InternalError"
  /faucet/status:
    get:
      tags: [Faucet]
      summary: Get faucet status
      description: Returns the current faucet status including available balance and rate limit info.
      operationId: getFaucetStatus
      responses:
        "200":
          description: Faucet status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    description: Whether the faucet is currently operational
                  balance:
                    type: string
                    description: Remaining faucet balance in uclaw
                  rate_limit_seconds:
                    type: integer
                    description: Cooldown period per address in seconds
              example:
                available: true
                balance: "999000000"
                rate_limit_seconds: 60
        "500":
          $ref: "#/components/responses/InternalError"

# ════════════════════════════════════════════════════════════════════════
#  Components
# ════════════════════════════════════════════════════════════════════════
components:
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
        message:
          type: string

    # ── Chain ──
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: "internal error"
            code: 13
            message: "an unexpected error occurred"